+++ loaders/csv/util.go
@@ -6,7 +6,7 @@ import (
 	"regexp"
 
 	"github.com/thalesfsp/customerror"
-	"github.com/thalesfsp/etler/v2/internal/shared"
+	"github.com/thalesfsp/etler/v3/internal/shared"
 )
 
 //////