+++ pipeline/interface.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"expvar"
 
-	"github.com/thalesfsp/etler/v2/internal/shared"
-	"github.com/thalesfsp/etler/v2/task"
+	"github.com/thalesfsp/etler/v3/internal/shared"
+	"github.com/thalesfsp/etler/v3/task"
 	"github.com/thalesfsp/status"
 )