+++ processors/storage/storage.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/thalesfsp/concurrentloop"
 	"github.com/thalesfsp/dal/storage"
-	"github.com/thalesfsp/etler/v2/internal/shared"
-	"github.com/thalesfsp/etler/v2/processor"
+	"github.com/thalesfsp/etler/v3/internal/shared"
+	"github.com/thalesfsp/etler/v3/processor"
 	"github.com/thalesfsp/params/create"
 	"github.com/thalesfsp/validation"
 )