+++ processors/storage/storage_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/thalesfsp/dal/memory"
-	"github.com/thalesfsp/etler/v2/processor"
+	"github.com/thalesfsp/etler/v3/processor"
 	"github.com/thalesfsp/params/list"
 )